added skip logic
